split of Runtime.h from Utils.h
