+++ enzyme/Enzyme/DebugUtils.h
@@ -1,0 +1,55 @@
+#ifndef ENZYME_DEBUG_UTILS_H
+#define ENZYME_DEBUG_UTILS_H
+
+#include <functional>
+#include <set>
+
+#include "llvm/ADT/SetVector.h"
+#include "llvm/ADT/SmallPtrSet.h"
+#include "llvm/IR/Value.h"
+
+#include "llvm/Support/raw_ostream.h"
+
+/// Output a set as a string
+template <typename T>
+static inline std::string to_string(const std::set<T> &us) {
+  std::string s = "{";
+  for (const auto &y : us)
+    s += std::to_string(y) + ",";
+  return s + "}";
+}
+
+/// Print a map, optionally with a shouldPrint function
+/// to decide to print a given value
+template <typename T, typename N>
+static inline void dumpMap(
+    const llvm::ValueMap<T, N> &o,
+    std::function<bool(const llvm::Value *)> shouldPrint = [](T) {
+      return true;
+    }) {
+  llvm::errs() << "<begin dump>\n";
+  for (auto a : o) {
+    if (shouldPrint(a.first))
+      llvm::errs() << "key=" << *a.first << " val=" << *a.second << "\n";
+  }
+  llvm::errs() << "</end dump>\n";
+}
+
+/// Print a set
+template <typename T>
+static inline void dumpSet(const llvm::SmallPtrSetImpl<T *> &o) {
+  llvm::errs() << "<begin dump>\n";
+  for (auto a : o)
+    llvm::errs() << *a << "\n";
+  llvm::errs() << "</end dump>\n";
+}
+
+template <typename T>
+static inline void dumpSet(const llvm::SetVector<T *> &o) {
+  llvm::errs() << "<begin dump>\n";
+  for (auto a : o)
+    llvm::errs() << *a << "\n";
+  llvm::errs() << "</end dump>\n";
+}
+
+#endif
