+++ enzyme/Enzyme/Enums.h
@@ -1,0 +1,258 @@
+#ifndef ENZYME_ENUMS_H
+#define ENZYME_ENUMS_H
+
+#include <set>
+#include <string>
+
+#include "llvm/IR/Type.h"
+#include "llvm/Support/raw_ostream.h"
+
+/// Potential return type of generated functions
+enum class ReturnType {
+  /// Return is a struct of all args and the original return
+  ArgsWithReturn,
+  /// Return is a struct of all args and two of the original return
+  ArgsWithTwoReturns,
+  /// Return is a struct of all args
+  Args,
+  /// Return is a tape type and the original return
+  TapeAndReturn,
+  /// Return is a tape type and the two of the original return
+  TapeAndTwoReturns,
+  /// Return is a tape type
+  Tape,
+  TwoReturns,
+  Return,
+  Void,
+};
+
+/// Convert ReturnType to a string
+static inline std::string to_string(ReturnType t) {
+  switch (t) {
+  case ReturnType::ArgsWithReturn:
+    return "ArgsWithReturn";
+  case ReturnType::ArgsWithTwoReturns:
+    return "ArgsWithTwoReturns";
+  case ReturnType::Args:
+    return "Args";
+  case ReturnType::TapeAndReturn:
+    return "TapeAndReturn";
+  case ReturnType::TapeAndTwoReturns:
+    return "TapeAndTwoReturns";
+  case ReturnType::Tape:
+    return "Tape";
+  case ReturnType::TwoReturns:
+    return "TwoReturns";
+  case ReturnType::Return:
+    return "Return";
+  case ReturnType::Void:
+    return "Void";
+  }
+  llvm_unreachable("illegal ReturnType");
+}
+
+/// Potential differentiable argument classifications
+enum class DIFFE_TYPE {
+  OUT_DIFF = 0,  // add differential to an output struct
+  DUP_ARG = 1,   // duplicate the argument and store differential inside
+  CONSTANT = 2,  // no differential
+  DUP_NONEED = 3 // duplicate this argument and store differential inside, but
+                 // don't need the forward
+};
+
+/// Convert DIFFE_TYPE to a string
+static inline std::string to_string(DIFFE_TYPE t) {
+  switch (t) {
+  case DIFFE_TYPE::OUT_DIFF:
+    return "OUT_DIFF";
+  case DIFFE_TYPE::CONSTANT:
+    return "CONSTANT";
+  case DIFFE_TYPE::DUP_ARG:
+    return "DUP_ARG";
+  case DIFFE_TYPE::DUP_NONEED:
+    return "DUP_NONEED";
+  default:
+    assert(0 && "illegal diffetype");
+    return "";
+  }
+}
+
+enum class BATCH_TYPE {
+  SCALAR = 0,
+  VECTOR = 1,
+};
+
+enum class DerivativeMode {
+  ForwardMode = 0,
+  ReverseModePrimal = 1,
+  ReverseModeGradient = 2,
+  ReverseModeCombined = 3,
+  ForwardModeSplit = 4,
+};
+
+static inline std::string to_string(DerivativeMode mode) {
+  switch (mode) {
+  case DerivativeMode::ForwardMode:
+    return "ForwardMode";
+  case DerivativeMode::ForwardModeSplit:
+    return "ForwardModeSplit";
+  case DerivativeMode::ReverseModePrimal:
+    return "ReverseModePrimal";
+  case DerivativeMode::ReverseModeGradient:
+    return "ReverseModeGradient";
+  case DerivativeMode::ReverseModeCombined:
+    return "ReverseModeCombined";
+  }
+  llvm_unreachable("illegal derivative mode");
+}
+
+enum class ProbProgMode {
+  Trace = 0,
+  Condition = 1,
+};
+
+static inline std::string to_string(ProbProgMode mode) {
+  switch (mode) {
+  case ProbProgMode::Trace:
+    return "Trace";
+  case ProbProgMode::Condition:
+    return "Condition";
+  }
+  llvm_unreachable("illegal probabalistic programming mode");
+}
+
+/// Classification of value as an original program
+/// variable, a derivative variable, neither, or both.
+/// This type is used both in differential use analysis
+/// and to describe argument bundles.
+enum class ValueType {
+  // A value that is neither a value in the original
+  // program, nor the derivative.
+  None = 0,
+  // The original program value
+  Primal = 1,
+  // The derivative value
+  Shadow = 2,
+  // Both the original program value and the shadow.
+  Both = Primal | Shadow,
+};
+
+using UsageKey = std::pair<const llvm::Value *, ValueType>;
+
+static inline std::string to_string(ValueType mode) {
+  switch (mode) {
+  case ValueType::None:
+    return "None";
+  case ValueType::Primal:
+    return "Primal";
+  case ValueType::Shadow:
+    return "Shadow";
+  case ValueType::Both:
+    return "Both";
+  }
+  llvm_unreachable("illegal valuetype");
+}
+
+/// Attempt to automatically detect the differentiable
+/// classification based off of a given type
+static inline DIFFE_TYPE whatType(llvm::Type *arg, DerivativeMode mode,
+                                  bool integersAreConstant,
+                                  std::set<llvm::Type *> &seen) {
+  assert(arg);
+  if (seen.find(arg) != seen.end())
+    return DIFFE_TYPE::CONSTANT;
+  seen.insert(arg);
+
+  if (arg->isVoidTy() || arg->isEmptyTy()) {
+    return DIFFE_TYPE::CONSTANT;
+  }
+
+  if (arg->isPointerTy()) {
+#if LLVM_VERSION_MAJOR >= 15
+    if (!arg->getContext().supportsTypedPointers()) {
+      return DIFFE_TYPE::DUP_ARG;
+    }
+#endif
+    switch (whatType(arg->getPointerElementType(), mode, integersAreConstant,
+                     seen)) {
+    case DIFFE_TYPE::OUT_DIFF:
+      return DIFFE_TYPE::DUP_ARG;
+    case DIFFE_TYPE::CONSTANT:
+      return DIFFE_TYPE::CONSTANT;
+    case DIFFE_TYPE::DUP_ARG:
+      return DIFFE_TYPE::DUP_ARG;
+    case DIFFE_TYPE::DUP_NONEED:
+      llvm_unreachable("impossible case");
+    }
+    assert(arg);
+    llvm::errs() << "arg: " << *arg << "\n";
+    assert(0 && "Cannot handle type0");
+    return DIFFE_TYPE::CONSTANT;
+  } else if (arg->isArrayTy()) {
+    return whatType(llvm::cast<llvm::ArrayType>(arg)->getElementType(), mode,
+                    integersAreConstant, seen);
+  } else if (arg->isStructTy()) {
+    auto st = llvm::cast<llvm::StructType>(arg);
+    if (st->getNumElements() == 0)
+      return DIFFE_TYPE::CONSTANT;
+
+    auto ty = DIFFE_TYPE::CONSTANT;
+    for (unsigned i = 0; i < st->getNumElements(); ++i) {
+      auto midTy =
+          whatType(st->getElementType(i), mode, integersAreConstant, seen);
+      switch (midTy) {
+      case DIFFE_TYPE::OUT_DIFF:
+        switch (ty) {
+        case DIFFE_TYPE::OUT_DIFF:
+        case DIFFE_TYPE::CONSTANT:
+          ty = DIFFE_TYPE::OUT_DIFF;
+          break;
+        case DIFFE_TYPE::DUP_ARG:
+          ty = DIFFE_TYPE::DUP_ARG;
+          return ty;
+        case DIFFE_TYPE::DUP_NONEED:
+          llvm_unreachable("impossible case");
+        }
+        break;
+      case DIFFE_TYPE::CONSTANT:
+        switch (ty) {
+        case DIFFE_TYPE::OUT_DIFF:
+          ty = DIFFE_TYPE::OUT_DIFF;
+          break;
+        case DIFFE_TYPE::CONSTANT:
+          break;
+        case DIFFE_TYPE::DUP_ARG:
+          ty = DIFFE_TYPE::DUP_ARG;
+          return ty;
+        case DIFFE_TYPE::DUP_NONEED:
+          llvm_unreachable("impossible case");
+        }
+        break;
+      case DIFFE_TYPE::DUP_ARG:
+        return DIFFE_TYPE::DUP_ARG;
+      case DIFFE_TYPE::DUP_NONEED:
+        llvm_unreachable("impossible case");
+      }
+    }
+    return ty;
+  } else if (arg->isIntOrIntVectorTy() || arg->isFunctionTy()) {
+    return integersAreConstant ? DIFFE_TYPE::CONSTANT : DIFFE_TYPE::DUP_ARG;
+  } else if (arg->isFPOrFPVectorTy()) {
+    return (mode == DerivativeMode::ForwardMode ||
+            mode == DerivativeMode::ForwardModeSplit)
+               ? DIFFE_TYPE::DUP_ARG
+               : DIFFE_TYPE::OUT_DIFF;
+  } else {
+    assert(arg);
+    llvm::errs() << "arg: " << *arg << "\n";
+    assert(0 && "Cannot handle type");
+    return DIFFE_TYPE::CONSTANT;
+  }
+}
+
+static inline DIFFE_TYPE whatType(llvm::Type *arg, DerivativeMode mode) {
+  std::set<llvm::Type *> seen;
+  return whatType(arg, mode, /*intconst*/ true, seen);
+}
+
+#endif
