+++ enzyme/Enzyme/ErrorHandling.h
@@ -1,0 +1,79 @@
+#ifndef ENZYME_ERROR_HANDLING_H
+#define ENZYME_ERROR_HANDLING_H
+
+#include <functional>
+#include <string>
+
+#include "llvm/IR/BasicBlock.h"
+#include "llvm/IR/DiagnosticInfo.h"
+#include "llvm/IR/Function.h"
+#include "llvm/IR/Instruction.h"
+
+#include "Runtime.h"
+#include "llvm/Support/ErrorHandling.h"
+#include "llvm/Support/raw_ostream.h"
+
+extern "C" {
+
+#include "llvm-c/Core.h"
+extern llvm::cl::opt<bool> EnzymePrintPerf;
+}
+
+template <typename... Args>
+void EmitWarning(llvm::StringRef RemarkName,
+                 const llvm::DiagnosticLocation &Loc,
+                 const llvm::BasicBlock *BB, const Args &...args) {
+
+  llvm::LLVMContext &Ctx = BB->getContext();
+  if (Ctx.getDiagHandlerPtr()->isPassedOptRemarkEnabled("enzyme")) {
+    std::string str;
+    llvm::raw_string_ostream ss(str);
+    (ss << ... << args);
+    auto R = llvm::OptimizationRemark("enzyme", RemarkName, Loc, BB)
+             << ss.str();
+    Ctx.diagnose(R);
+  }
+
+  if (EnzymePrintPerf)
+    (llvm::errs() << ... << args) << "\n";
+}
+
+template <typename... Args>
+void EmitWarning(llvm::StringRef RemarkName, const llvm::Instruction &I,
+                 const Args &...args) {
+  EmitWarning(RemarkName, I.getDebugLoc(), I.getParent(), args...);
+}
+
+template <typename... Args>
+void EmitWarning(llvm::StringRef RemarkName, const llvm::Function &F,
+                 const Args &...args) {
+  llvm::LLVMContext &Ctx = F.getContext();
+  if (Ctx.getDiagHandlerPtr()->isPassedOptRemarkEnabled("enzyme")) {
+    std::string str;
+    llvm::raw_string_ostream ss(str);
+    (ss << ... << args);
+    auto R = llvm::OptimizationRemark("enzyme", RemarkName, &F) << ss.str();
+    Ctx.diagnose(R);
+  }
+  if (EnzymePrintPerf)
+    (llvm::errs() << ... << args) << "\n";
+}
+
+class EnzymeFailure final : public llvm::DiagnosticInfoUnsupported {
+public:
+  EnzymeFailure(llvm::Twine Msg, const llvm::DiagnosticLocation &Loc,
+                const llvm::Instruction *CodeRegion);
+};
+
+template <typename... Args>
+void EmitFailure(llvm::StringRef RemarkName,
+                 const llvm::DiagnosticLocation &Loc,
+                 const llvm::Instruction *CodeRegion, Args &...args) {
+  std::string *str = new std::string();
+  llvm::raw_string_ostream ss(*str);
+  (ss << ... << args);
+  CodeRegion->getContext().diagnose(
+      (EnzymeFailure(llvm::Twine("Enzyme: ") + ss.str(), Loc, CodeRegion)));
+}
+
+#endif
