+++ enzyme/Enzyme/MapUtils.h
@@ -1,0 +1,57 @@
+#ifndef ENZYME_MAP_UTILS_H
+#define ENZYME_MAP_UTILS_H
+
+#include <map>
+
+#include "llvm/ADT/DenseMapInfo.h"
+#include "llvm/IR/Value.h"
+
+class AssertingReplacingVH final : public llvm::CallbackVH {
+public:
+  AssertingReplacingVH() = default;
+
+  AssertingReplacingVH(llvm::Value *new_value) { setValPtr(new_value); }
+
+  void deleted() override final {
+    assert(0 && "attempted to delete value with remaining handle use");
+    llvm_unreachable("attempted to delete value with remaining handle use");
+  }
+
+  void allUsesReplacedWith(llvm::Value *new_value) override final {
+    setValPtr(new_value);
+  }
+  virtual ~AssertingReplacingVH() {}
+};
+
+/// Insert into a map
+template <typename K, typename V>
+static inline typename std::map<K, V>::iterator
+insert_or_assign(std::map<K, V> &map, K &key, V &&val) {
+  auto found = map.find(key);
+  if (found != map.end()) {
+    map.erase(found);
+  }
+  return map.emplace(key, val).first;
+}
+
+/// Insert into a map
+template <typename K, typename V>
+static inline typename std::map<K, V>::iterator
+insert_or_assign2(std::map<K, V> &map, K key, V val) {
+  auto found = map.find(key);
+  if (found != map.end()) {
+    map.erase(found);
+  }
+  return map.emplace(key, val).first;
+}
+
+template <typename K, typename V>
+static inline V *findInMap(std::map<K, V> &map, K key) {
+  auto found = map.find(key);
+  if (found == map.end())
+    return nullptr;
+  V *val = &found->second;
+  return val;
+}
+
+#endif
